update(propagation): Changed P.2108 parameter type to clutter_scenario.
The type parameter expects the StationType to check if the clutter
scenario is terrestrial or spacial. This is prone to errors
if the study is between IMT and another station that is not FSS_ES
or FSS_SS.
Now the user specifies if it's "terrestrial" path (P.2108 sec 3.2)
or "spacial" path (P.2108 sec 3.3).